% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAIFinalise.R
\name{CAIfinalise}
\alias{CAIfinalise}
\title{Arrange alignment and save CAI ggplot chart.}
\source{
BBCplot (function) reworked.
}
\usage{
CAIfinalise(plot_name = last_plot(), source = "CA&I  ©",
  filename = "unnamed_plot.png", filepath = getwd(), size = "full",
  width = 640, height = 450, unit = "px",
  logo_path = "https://bit.ly/2Z5JeU3")
}
\arguments{
\item{plot_name}{The variable name of the plot you have created that you want to format and save}

\item{source}{The text you want to come after the text 'Source:' in the bottom left hand side of your side.
"None" for no source}

\item{filename}{Exact name of file that you want to save to}

\item{filepath}{Exact path to the directory you want to save to}

\item{size}{Specify the size - "full", "half", "long", or "manual", which references height and width}

\item{width}{Width in pixels that you want to save your chart to - defaults to 640}

\item{height}{Height in pixels that you want to save your chart to - defaults to 450}

\item{unit}{Default is pixels (px), but accepts cm}

\item{logo_path}{File path for the logo image you want to use in the right hand side of your chart,
which needs to be a PNG file - defaults to CAI blocks image that sits within the data
folder of your package}
}
\value{
(Invisibly) an updated ggplot object.
}
\description{
Function for finalising a graphic - resize, add logo and source, then save.
}
\details{
Running this function will save your plot with the correct guidelines for publication for a CAI graphic.
It will left align your title, subtitle and source, add the CAI logo at the bottom right and save it to your specified location.
}
\keyword{finalise_plot}
