% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanNames.R
\name{CleanNames}
\alias{CleanNames}
\title{Clean column names in strings or data.frames}
\usage{
CleanNames(string, unique = FALSE)
}
\arguments{
\item{string}{Either a data.frame or string for cleaning.}

\item{unique}{If False (default), allows multiple occurences of each name.}
}
\value{
No return or comments, unless package doesn't exit.
}
\description{
Rationalise each of the strings, removing spaces, changing symbols and removing duplication.
}
\details{
Sourced from https://drdoane.com/clean-consistent-column-names/
}
