% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstallLoad.R
\name{InstallLoad}
\alias{InstallLoad}
\title{Mutliple load of libraries.}
\usage{
InstallLoad(packages)
}
\arguments{
\item{packages}{One or more libraries to be loaded.}
}
\value{
No return or comments, unless package doesn't exit.
}
\description{
Install each package in turn, and load the libraries
}
\details{
Function to load (if necessary) and install multiple packages at once
Sourced from https://www.listendata.com/2018/12/install-load-multiple-r-packages.html
Only works with CRAN packages. Dev packages need to be loaded seperately
}
\examples{
packages <- c("extrafont", "scales")
InstallLoad(packages)
}
