% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstallLoad.R
\name{InstallLoad}
\alias{InstallLoad}
\title{Mutliple load of libraries.}
\usage{
InstallLoad(packages = "", supress = TRUE, default = TRUE,
  verbose = FALSE)
}
\arguments{
\item{packages}{One or more libraries to be loaded.}

\item{supress}{If TRUE (default), package startup messages will be suppressed.}

\item{default}{Loads a default set of libraries,
e.e. Tidyverse, CAIplot, grid, extrafont, scales}

\item{verbose}{Provide verbose messages as packages load}
}
\value{
No return or comments, unless package doesn't exit.
}
\description{
Install each package in turn, and load the libraries
}
\details{
Function to install packages (if necessary) and load them at once
Sourced from https://www.listendata.com/2018/12/install-load-multiple-r-packages.html
Only works with CRAN packages. Dev packages need to be loaded seperately
}
